CategoryDetail ViewModel File
