+++ TravelDiscovery/Discover/CategoryDetail/View/CategoryDetailView.swift
@@ -29,7 +29,6 @@ struct CategoryDetailView: View {
             .padding()
             .background(Color.black)
             .cornerRadius(8)
-            
         } else {
             ZStack {
                 if !vm.errorMessage.isEmpty {
@@ -44,8 +43,6 @@ struct CategoryDetailView: View {
                     ForEach(vm.places, id: \.self) { place in
                         VStack(alignment: .leading, spacing: 0) {
                             KFImage(URL(string: place.thumbnail))
-//                            WebImage(url: URL(string: place.thumbnail))
-//                            AsyncImage(url: URL(string: place.thumbnail))
                                 .resizable()
                                 .scaledToFill()
                             Text(place.name)
@@ -59,40 +56,6 @@ struct CategoryDetailView: View {
             }
             .navigationBarTitle(name, displayMode: .inline)
         }
-    }
-}
-
-class CategoryDetailsViewModel: ObservableObject {
-    @Published var isLoading = true
-    @Published var places = [Place]()
-    @Published var errorMessage = ""
-    init(name: String) {
-        let urlString = "https://travel.letsbuildthatapp.com/travel_discovery/category?name=\(name.lowercased())".addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) ?? ""
-        guard let url = URL(string: urlString)
-        else {
-            self.isLoading = false
-            return
-        }
-        
-        URLSession.shared.dataTask(with: url) { data, response, error in
-
-            DispatchQueue.main.asyncAfter(deadline: .now() + 2) {
-                if let statusCode = (response as? HTTPURLResponse)?.statusCode, statusCode >= 400 {
-                    self.isLoading = false
-                    self.errorMessage = "Bad status: \(statusCode)"
-                    return
-                }
-
-                guard let data = data else { return }
-                do {
-                    self.places = try JSONDecoder().decode([Place].self, from: data)
-                } catch {
-                    print("Failed to decode JSON: ", error)
-                    self.errorMessage = error.localizedDescription
-                }
-                self.isLoading = false
-            }
-        }.resume()
     }
 }
